move controllers test under controllers folder
