+++ tests/controllers/userController/createACSUser.test.ts
@@ -4,7 +4,7 @@
  *---------------------------------------------------------------------------------------------*/
 
 // eslint-disable-next-line @typescript-eslint/triple-slash-reference
-/// <reference path="../../node_modules/@types/jest/index.d.ts" />
+/// <reference path="../../../node_modules/@types/jest/index.d.ts" />
 
 import {
   mockAcsUserId,
@@ -13,11 +13,11 @@ import {
   mockIdentityMapping,
   mockResponse,
   mockRequest
-} from '../utils/mockData';
-import { createACSUser } from '../../src/controllers/userController';
-import * as acsService from '../../src/services/acsService';
-import * as aadService from '../../src/services/aadService';
-import * as graphService from '../../src/services/graphService';
+} from '../../utils/mockData';
+import { createACSUser } from '../../../src/controllers/userController';
+import * as acsService from '../../../src/services/acsService';
+import * as aadService from '../../../src/services/aadService';
+import * as graphService from '../../../src/services/graphService';
 
 let exchangeAADTokenViaOBOSpy: jest.SpyInstance;
 let getACSUserIdSpy: jest.SpyInstance;
